+++ RELEASE_NOTES.md
@@ -5,6 +5,7 @@ This is the offical release of vpinfe v1.1.20.
 
 ### What's New
 - Add global "-tableini" masking override for vpx table launching. 
+- Add a live view of the vpinball actual launching command
 
 ### Fixes
 None